# Perf(zstd): Improve 'matchLen' performance by vector instructions.