--- a/zstd/_generate/go.mod
+++ b/zstd/_generate/go.mod
@@ -6,7 +6,6 @@ require (
 	github.com/klauspost/compress v1.15.15
 	github.com/mmcloughlin/avo v0.5.0
 	golang.org/x/tools v0.6.0 // indirect
-	golang.org/x/xerrors v0.0.0-20220907171357-04be3eba64a2 // indirect
 )
 
 replace github.com/klauspost/compress => ../..