… zstd/_generate/go.sum
… zstd/matchlen_amd64.s