@zzzzwc 2023-06-07
And I will write the code to make it work fine on amd64 cpu that does not support avx2