@klauspost 2023-06-07
Current code is a big speed regression. Testing on [CSV input](https://files.klauspost.com/compress/nyc-taxi-data-10M.csv.zst):

Before/after (3 runs):
```
file	out	level	insize	outsize	millis	mb/s
nyc-taxi-data-10M.csv	zskp	1	3325605752	641319332	8373	378.75
nyc-taxi-data-10M.csv	zskp	1	3325605752	641319332	8342	380.19
nyc-taxi-data-10M.csv	zskp	1	3325605752	641319332	8346	379.98
file	out	level	insize	outsize	millis	mb/s
nyc-taxi-data-10M.csv	zskp	1	3325605752	641319332	21627	146.64
nyc-taxi-data-10M.csv	zskp	1	3325605752	641319332	21093	150.35
nyc-taxi-data-10M.csv	zskp	1	3325605752	641319332	21799	145.49
```