@klauspost 2023-06-07
Adding VZEROUPPER fixes that:

```
file	out	level	insize	outsize	millis	mb/s
nyc-taxi-data-10M.csv	zskp	2	3325605752	588976126	9472	334.82
nyc-taxi-data-10M.csv	zskp	2	3325605752	588976126	9524	332.98
nyc-taxi-data-10M.csv	zskp	2	3325605752	588976126	9475	334.70
```

So if you can fix the issues, this could be feasible.