@zzzzwc 2023-06-07
> Current code is a big speed regression. Testing on [CSV input](https://files.klauspost.com/compress/nyc-taxi-data-10M.csv.zst):
> 
> Before/after (3 runs):
> 
> ```
> file	out	level	insize	outsize	millis	mb/s
> nyc-taxi-data-10M.csv	zskp	1	3325605752	641319332	8373	378.75
> nyc-taxi-data-10M.csv	zskp	1	3325605752	641319332	8342	380.19
> nyc-taxi-data-10M.csv	zskp	1	3325605752	641319332	8346	379.98
> file	out	level	insize	outsize	millis	mb/s
> nyc-taxi-data-10M.csv	zskp	1	3325605752	641319332	21627	146.64
> nyc-taxi-data-10M.csv	zskp	1	3325605752	641319332	21093	150.35
> nyc-taxi-data-10M.csv	zskp	1	3325605752	641319332	21799	145.49
> ```
> 
> Level 2 (default)
> 
> ```
> file	out	level	insize	outsize	millis	mb/s
> nyc-taxi-data-10M.csv	zskp	2	3325605752	588976126	10543	300.82
> nyc-taxi-data-10M.csv	zskp	2	3325605752	588976126	10621	298.60
> nyc-taxi-data-10M.csv	zskp	2	3325605752	588976126	10634	298.22
> file	out	level	insize	outsize	millis	mb/s
> nyc-taxi-data-10M.csv	zskp	2	3325605752	588976126	23413	135.46
> nyc-taxi-data-10M.csv	zskp	2	3325605752	588976126	23136	137.08
> nyc-taxi-data-10M.csv	zskp	2	3325605752	588976126	23133	137.10
> ```

Is there code for the above benchmark tests in this repo? I did not find it