@klauspost 2023-06-09
I copied over the [matchlen assembly from s2](https://github.com/klauspost/compress/blob/master/s2/encodeblock_amd64.s#L18868), and it is faster than AVX2 based - and it doesn't use anything beyond standard amd64:

```
file	out	level	insize	outsize	millis	mb/s
nyc-taxi-data-10M.csv	zskp	1	3325605752	641319332	8171	388.12
nyc-taxi-data-10M.csv	zskp	1	3325605752	641319332	8104	391.33
nyc-taxi-data-10M.csv	zskp	1	3325605752	641319332	8109	391.08
file	out	level	insize	outsize	millis	mb/s
nyc-taxi-data-10M.csv	zskp	2	3325605752	588976126	10077	314.73
nyc-taxi-data-10M.csv	zskp	2	3325605752	588976126	10259	309.13
nyc-taxi-data-10M.csv	zskp	2	3325605752	588976126	10052	315.50
```