@zzzzwc 2023-06-12
the cost of matchLen looks like:

<img width="1657" alt="image" src="https://github.com/klauspost/compress/assets/94814774/ebc398e5-0745-405d-8a76-411fb9c2b5c5">
x is matched length

y is cost time

the green line is the normal way and the blue line is the vectorized way

I'm trying to reduce the external cost `20` and `2`