@klauspost 2023-06-12
@zzzzwc I am not sure how you got the equations, but they show the problem fine. You show a crossover at 320 bytes. That is *incredibly* rare. Most matches are much less than 100 bytes.

`VPMOVMSKB` is a vector -> GPR register transfer. This incurs a 4-7 cycle latency depending on what platform you measure. Also having to add `NOT` is an additional 1 cycle. But I presume that is what your "20" accounts for.

This means the penalty for a 4-8 (and possibly up to 16) byte match is much higher. Also add that when loading 32 bytes you are much more likely to cross a 64byte cache line which incurs a minor load penalty.

However, I would expect branch prediction to be better, since it is very likely that only one loop will be called. However, it is not enough that it gives an overall improvement. 

I think the crossover is lower than what your graph indicates. But you are optimizing the wrong thing. We don't want *long* matches to be faster. We want *short* matches to be faster, since they are the main bulk, and fast matches are not a problem for overall performance.

I tried adding it to S2 assembly https://github.com/klauspost/compress/pull/825 with appropriate fallback - and again, it is not faster than the BSF code, so I left it disabled. I left the code in, so you can see how it is integrated.

It is important that you test on real data.